refactor(console): move 2 api component to parent component directory
