+++ gravitee-apim-console-webui/src/components/api-subscription-edit-push-config-dialog/api-subscription-edit-push-config-dialog.component.scss
