+++ gravitee-apim-console-webui/src/components/api-subscription-edit-push-config/api-subscription-edit-push-config.component.scss
